requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
certifi==2019.11.28
chardet==3.0.4
cryptography==2.8
jsonschema==3.2.0
numpy==1.19.4
keyring==18.0.1
pandas==1.1.5
python-telegram-bot==13.1
pytz==2019.3
tornado==6.1
python-decouple==3.3
simplejson==3.16.0
soupsieve==2.1
